Changed label loading within training

+++ cxr_ssgan.py
@@ -353,7 +353,8 @@ for epoch in range(num_epochs):
         labeled_data, labels, label_mask = data
         labeled_data = _to_var(labeled_data).float()
         for i, label in enumerate(labels):
-            labels[i] = torch.FloatTensor(label)
+            labels[i] = torch.LongTensor(label)
+
         labels = _to_var(labels)
 
         labels = torch.cat((labels, _to_var(torch.ones(labels.shape[0], 1))), 1)
